% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_pinyin.R
\name{list_phrases}
\alias{list_phrases}
\title{List Custom Polyphone Phrases}
\usage{
list_phrases()
}
\value{
A data frame with columns \code{phrase}, \code{tone}, and \code{marks}.
}
\description{
List Custom Polyphone Phrases
}
\examples{
list_phrases()
}
